Seperated structs into views
