+++ nVolve/Views/EventCardView.swift
@@ -1,5 +1,5 @@
 //
-//  CardView.swift
+//  EventCard.swift
 //  nVolve
 //
 //  Created by Abdalla Abdelmagid on 11/11/24.
@@ -7,17 +7,17 @@
 
 import SwiftUI
 
-struct CardView: View {
-    var imagePath: String?
-    var title: String
-    var time: String
-    var room: String
+struct EventCard: View {
+    let event: InvolvedEvent?
+    let date: String
+    let imagePath: String?
+    @Binding var showEvent: Bool
 
     var body: some View {
         VStack(alignment: .leading, spacing: 4) {
             if let imagePath = imagePath {
-                AsyncImage(url: URL(string: imagePath)) {
-                    image in image
+                AsyncImage(url: URL(string: imagePath)) { image in
+                    image
                         .resizable()
                         .scaledToFit()
                         .frame(height: 120)
@@ -28,21 +28,21 @@ struct CardView: View {
                                 .stroke(Color.purple, lineWidth: 1)
                         )
                 } placeholder: {
-                    ProgressView("Boof")
+                    ProgressView("Loading...")
                 }
             }
 
             VStack(alignment: .leading, spacing: 3) {
-                Text(title)
+                Text(event?.eventName ?? "Event")
                     .font(.headline)
                     .fontWeight(.semibold)
                     .lineLimit(1)
 
-                Text("\(time) / \(room)")
+                Text("\(date) / \(event?.eventLocation ?? "Room")")
                     .font(.caption)
                     .bold()
 
-                Text("Description")
+                Text(event?.eventDescription ?? "Description")
                     .font(.caption)
                     .lineLimit(2)
             }
@@ -60,16 +60,31 @@ struct CardView: View {
                 .stroke(Color.yellow, lineWidth: 1)
         )
         .shadow(color: Color.gray.opacity(0.2), radius: 2, x: 0, y: 1)
+        .onTapGesture {
+            showEvent = true
+        }
+        .fullScreenCover(isPresented: $showEvent) {
+            EventInfoView(
+                showEvent: $showEvent,
+                title: event?.eventName ?? "Event Title",
+                time: date,
+                room: event?.eventLocation ?? "Room 204",
+                description: event?.eventDescription ?? "Description",
+                eventLat: String(format: "%.6f", event?.latitude ?? 0.0),
+                eventLng: String(format: "%.6f", event?.longitude ?? 0.0),
+                perks: event?.perks ?? ["Free Food", "Networking"]
+            )
+        }
     }
 }
 
-struct Card_Previews: PreviewProvider {
+struct EventCard_Previews: PreviewProvider {
     static var previews: some View {
-        CardView(
+        EventCard(
+            event: nil,
+            date: "10:00 AM",
             imagePath: "https://via.placeholder.com/300",
-            title: "Event Title",
-            time: "10:00 AM",
-            room: "Room 204"
+            showEvent: .constant(false)
         )
         .previewLayout(.sizeThatFits)
         .padding()
